fix(benchmark): correct CUDA autolykos DAG index overflow in prehash kernels (untested)
The benchmark autolykos prehash kernels (autolykos_v2_prehash_v1.cu,
autolykos_v2_mhssamadani_prehash.cu) carry the same hashes[tid*32+31] 32-bit
overflow as cuda/dag.cuh, and the benchmark runs with period 146,488,965 (> the
134,217,728 element / 4 GiB boundary), so they would write past the wrap point.
Benchmark-only (perf measurement does not gate correctness), but fixed for
consistency. Cast the byte index to uint64_t. Untested on NVIDIA hardware.
luminousmining/miner#159.

+++ sources/benchmark/cuda/autolykos_v2/autolykos_v2_mhssamadani_prehash.cu
@@ -210,7 +210,9 @@ void autolykos_v2_mhssamadi_init_prehash(
         {
             store64(&(((uint64_t*)hashes)[(tid + 1) * 4 - i - 1]), h[i]);
         }
-        ((uint8_t*)hashes)[tid * 32 + 31] = 0;
+        // (uint64_t) cast: tid*32 overflows 32-bit for tid >= 2^27 (4 GiB). Same bug as
+        // cuda/dag.cuh -- luminousmining/miner#159. Untested on NVIDIA hardware.
+        ((uint8_t*)hashes)[(uint64_t)tid * 32 + 31] = 0;
     }
 }
